Used Zoomable component for profile zoom image
