+++ src/components/post/FeedPost.tsx
@@ -15,20 +15,9 @@ import {
   useWindowDimensions,
 } from 'react-native'
 import { Blurhash } from 'react-native-blurhash'
-import {
-  Gesture,
-  GestureDetector,
-  PinchGestureHandler,
-  State,
-} from 'react-native-gesture-handler'
+import { Gesture, GestureDetector } from 'react-native-gesture-handler'
 import { PressableOpacity } from 'react-native-pressable-opacity'
-import Animated, {
-  runOnJS,
-  type SharedValue,
-  useAnimatedStyle,
-  useSharedValue,
-  withSpring,
-} from 'react-native-reanimated'
+import { runOnJS, type SharedValue, useSharedValue } from 'react-native-reanimated'
 import Carousel, { Pagination } from 'react-native-reanimated-carousel'
 import ImageComponent from 'src/components/ImageComponent'
 import AutolinkText from 'src/components/common/AutolinkText'
@@ -46,12 +35,6 @@ import { Button, Separator, Text, View, XStack, YStack, ZStack } from 'tamagui'
 import ReadMore from '../common/ReadMore'
 import VideoPlayer from './VideoPlayer'
 
-import { Image } from 'expo-image'
-import type {
-  GestureEvent,
-  HandlerStateChangeEvent,
-  PinchGestureHandlerEventPayload,
-} from 'react-native-gesture-handler'
 import { useBookmarkMutation } from 'src/hooks/mutations/useBookmarkMutation'
 import { useLikeMutation } from 'src/hooks/mutations/useLikeMutation'
 import type {
@@ -64,72 +47,7 @@ import type {
   Visibility,
 } from 'src/lib/api-types'
 import { PixelfedBottomSheetModal } from '../common/BottomSheets'
-
-const AnimatedFastImage = Animated.createAnimatedComponent(Image)
-
-const ZoomableImage = ({ source, placeholder, style }) => {
-  const scale = useSharedValue(1)
-  const savedScale = useSharedValue(1)
-  const translateX = useSharedValue(0)
-  const translateY = useSharedValue(0)
-  const originX = useSharedValue(0)
-  const originY = useSharedValue(0)
-
-  const onGestureEvent = (event: GestureEvent<PinchGestureHandlerEventPayload>) => {
-    const pinchScale = event.nativeEvent.scale
-    const nextScale = savedScale.value * pinchScale
-    const touchX = event.nativeEvent.focalX
-    const touchY = event.nativeEvent.focalY
-
-    if (scale.value === savedScale.value) {
-      originX.value = touchX
-      originY.value = touchY
-    }
-
-    const focalDeltaX = (touchX - originX.value) * (pinchScale - 1)
-    const focalDeltaY = (touchY - originY.value) * (pinchScale - 1)
-
-    scale.value = nextScale
-    translateX.value = focalDeltaX
-    translateY.value = focalDeltaY
-  }
-
-  const onHandlerStateChange = (
-    event: HandlerStateChangeEvent<PinchGestureHandlerEventPayload>
-  ) => {
-    if (event.nativeEvent.oldState === State.ACTIVE) {
-      savedScale.value = scale.value
-      scale.value = withSpring(1)
-      savedScale.value = 1
-      translateX.value = withSpring(0)
-      translateY.value = withSpring(0)
-    }
-  }
-
-  const animatedStyle = useAnimatedStyle(() => ({
-    transform: [
-      { translateX: translateX.value },
-      { translateY: translateY.value },
-      { scale: scale.value },
-    ],
-  }))
-
-  return (
-    <PinchGestureHandler
-      onGestureEvent={onGestureEvent}
-      onHandlerStateChange={onHandlerStateChange}
-    >
-      <Animated.View>
-        <AnimatedFastImage
-          source={source}
-          placeholder={placeholder}
-          style={[style, animatedStyle]}
-          contentFit={'cover'}
-        />
-      </Animated.View>
-    </PinchGestureHandler>
-  )
-}
+import ZoomableImage from '../common/ZoomableImage'
 
 const AVATAR_WIDTH = 45
